@@ -1,12 +1,12 @@
 import asyncio
 import os
 import re
+import signal
 import sys
 from glob import glob
 from os import sep
-import signal
 from time import time
-from typing import Optional, Tuple, Union, Any, Type, cast
+from typing import Any, Optional, Tuple, Type, Union, cast
 
 import discord
 import psutil
@@ -16,29 +16,35 @@ from discord.ext import commands
 from discord.utils import oauth_url
 from envparse import env
 
-from . import __version__
-from .classes.database import Database
-from .models.guild import Guild
-from .models.setting import Setting
-from .utils import log
-from .utils.constants import PERMISSIONS
-from .utils.context_menu import load_context_menu
-from .views.ExchangeGiftView import ExchangeGiftView
+from neonbot import __version__
+from neonbot.classes.database import Database
+from neonbot.models.flyff import FlyffModel
+from neonbot.models.guild import GuildModel
+from neonbot.models.setting import SettingModel
+from neonbot.utils import log
+from neonbot.utils.constants import PERMISSIONS
+from neonbot.utils.context_menu import load_context_menu
+from neonbot.views.ExchangeGiftView import ExchangeGiftView
 
 
 class NeonBot(commands.Bot):
     def __init__(self):
-        self.default_prefix = env.str("DEFAULT_PREFIX", default=".")
-        self.user_agent = f"NeonBot v{__version__}"
+        self.default_prefix = env.str('DEFAULT_PREFIX', default='.')
+        self.user_agent = f'NeonBot v{__version__}'
         self.loop = asyncio.get_event_loop()
-        super().__init__(intents=discord.Intents.all(), command_prefix=self.default_prefix,
-                         owner_ids=set(env.list("OWNER_IDS", default=[], subcast=int)))
+        self.executor = None
+        super().__init__(
+            intents=discord.Intents.all(),
+            command_prefix=self.default_prefix,
+            owner_ids=set(env.list('OWNER_IDS', default=[], subcast=int)),
+        )
 
         self.db = Database(self)
         self.app_info: Optional[discord.AppInfo] = None
         self.owner_guilds = env.list('OWNER_GUILD_IDS', default=[], subcast=int)
         self.session: Optional[ClientSession] = None
-        self.setting: Optional[Setting] = None
+        self.setting: Optional[SettingModel] = None
+        self.flyff_settings: Optional[FlyffModel] = None
         self.scheduler: Optional[AsyncIOScheduler] = None
         self.is_listeners_done = False
         self.is_player_cache_loaded = False
@@ -50,16 +56,15 @@ class NeonBot(commands.Bot):
 
         return (
             discord.Status[status],
-            discord.Activity(
-                name=activity_name, type=discord.ActivityType[activity_type]
-            ),
+            discord.Activity(name=activity_name, type=discord.ActivityType[activity_type]),
         )
 
     async def setup_hook(self):
         self.loop.add_signal_handler(signal.SIGTERM, lambda: asyncio.create_task(self.close()))
 
         await self.db.initialize()
-        self.setting = await Setting.get_instance()
+        self.setting = await SettingModel.get_instance()
+        self.flyff_settings = await FlyffModel.get_instance()
         self.status, self.activity = self.get_presence()
         self.session = ClientSession(timeout=ClientTimeout(total=30))
         self.scheduler = AsyncIOScheduler()
@@ -81,16 +86,19 @@ class NeonBot(commands.Bot):
 
     async def sync_command(self, guild: Optional[discord.Guild] = None):
         await self.tree.sync(guild=guild)
-        log.info(f"Command synced to: {guild or 'Global'}")
+        log.info(f'Command synced to: {guild or "Global"}')
 
     def start_listeners(self):
         if self.is_listeners_done:
             return
 
-        from .classes.panel import Panel
+        from neonbot.classes.panel import Panel
+        from neonbot.classes.flyff import Flyff
+
+        Flyff.start_listener()
 
         for guild in self.guilds:
-            server = Guild.get_instance(guild.id)
+            server = GuildModel.get_instance(guild.id)
 
             if server and not server.exchange_gift.finish and server.exchange_gift.message_id:
                 self.add_view(ExchangeGiftView(), message_id=server.exchange_gift.message_id)
@@ -103,7 +111,7 @@ class NeonBot(commands.Bot):
         if not env.bool('LOAD_PLAYER_CACHE', default=False) or self.is_player_cache_loaded:
             return
 
-        from .classes.player import Player
+        from neonbot.classes.player import Player
 
         for guild in self.guilds:
             if Player.has_cache(guild.id):
@@ -113,20 +121,20 @@ class NeonBot(commands.Bot):
         self.is_player_cache_loaded = True
 
     async def add_cogs(self):
-        files = sorted(glob(f"neonbot{sep}cogs{sep}[!_]*.py"))
-        extensions = [re.split(r"[{0}.]".format(re.escape(sep)), file)[-2] for file in files]
+        files = sorted(glob(f'neonbot{sep}cogs{sep}[!_]*.py'))
+        extensions = [re.split(r'[{0}.]'.format(re.escape(sep)), file)[-2] for file in files]
         start_time = time()
         process = psutil.Process(os.getpid())
 
         print(file=sys.stderr)
 
         for extension in extensions:
-            log.info(f"Loading {extension} cog... [{(process.memory_info().rss / 1024000):.2f} MB]")
-            await self.load_extension("neonbot.cogs." + extension)
+            log.info(f'Loading {extension} cog... [{(process.memory_info().rss / 1024000):.2f} MB]')
+            await self.load_extension('neonbot.cogs.' + extension)
 
         print(file=sys.stderr)
 
-        log.info(f"Loaded {len(extensions)} cogs after {(time() - start_time):.2f}s\n")
+        log.info(f'Loaded {len(extensions)} cogs after {(time() - start_time):.2f}s\n')
 
     async def fetch_app_info(self) -> None:
         if not self.app_info:
@@ -136,27 +144,26 @@ class NeonBot(commands.Bot):
         url = oauth_url(
             self.app_info.id,
             permissions=discord.Permissions(permissions=PERMISSIONS),
-            scopes=('bot', 'applications.commands')
+            scopes=('bot', 'applications.commands'),
         )
-        await message.channel.send(f"Bot invite link: {url}")
-        log.info(f"Sent an invite link to: {message.author}")
+        await message.channel.send(f'Bot invite link: {url}')
+        log.info(f'Sent an invite link to: {message.author}')
 
     async def update_presence(self):
-        setting = await Setting.get_instance()
+        setting = await SettingModel.get_instance()
 
         await self.change_presence(
-            activity=discord.Activity(
-                name=setting.activity_name,
-                type=discord.ActivityType[setting.activity_type]
-            ),
-            status=getattr(discord.Status, setting.status)
+            activity=discord.Activity(name=setting.activity_name, type=discord.ActivityType[setting.activity_type]),
+            status=getattr(discord.Status, setting.status),
         )
 
     async def send_response(self, interaction: discord.Interaction, *args, **kwargs):
         if not cast(discord.InteractionResponse, interaction.response).is_done():
             await cast(discord.InteractionResponse, interaction.response).send_message(*args, **kwargs)
-        elif cast(discord.InteractionResponse,
-                  interaction.response).type == discord.InteractionResponseType.deferred_message_update:
+        elif (
+            cast(discord.InteractionResponse, interaction.response).type
+            == discord.InteractionResponseType.deferred_message_update
+        ):
             await interaction.followup.send(*args, **kwargs)
         else:
             if 'view' not in kwargs:
@@ -173,20 +180,17 @@ class NeonBot(commands.Bot):
 
         try:
             await message.edit(**kwargs)
-        except discord.NotFound:
+        except:
             pass
 
     async def delete_message(self, *messages: Union[discord.Message, None]) -> None:
-        await asyncio.gather(
-            *[message.delete() for message in messages if message is not None],
-            return_exceptions=True
-        )
+        await asyncio.gather(*[message.delete() for message in messages if message is not None], return_exceptions=True)
 
     async def send_to_owner(self, *args: Any, **kwargs: Any) -> None:
         await self.get_user(self.app_info.owner.id).send(*args, **kwargs)
 
     async def close(self) -> None:
-        from .classes.player import Player
+        from neonbot.classes.player import Player
 
         if self.scheduler:
             log.info('Stopping scheduler...')
@@ -209,6 +213,8 @@ class NeonBot(commands.Bot):
         await super().start(*args, **kwargs)
 
     def run(self, *args, **kwargs):
+        self.executor = kwargs['executor']
+        del kwargs['executor']
         super().run(env.str('TOKEN'), *args, **kwargs)
 
     def _handle_ready(self) -> None:
