@@ -1,10 +1,11 @@
 import logging
 import os
 import shutil
+from concurrent.futures import ThreadPoolExecutor
 
 import i18n
-from envparse import env
 from dotenv import load_dotenv
+from envparse import env
 
 load_dotenv()
 env.read_envfile()
@@ -15,17 +16,22 @@ i18n.set('skip_locale_root_data', True)
 
 def main() -> None:
     from neonbot import bot
-    from neonbot.utils.constants import YOUTUBE_DOWNLOADS_DIR, PLAYER_CACHE_DIR
+    from neonbot.utils.constants import PLAYER_CACHE_DIR, YOUTUBE_DOWNLOADS_DIR
 
-    shutil.rmtree(YOUTUBE_DOWNLOADS_DIR, ignore_errors=True)
+    if env.bool('YTDL_AUTO_CLEAR_DOWNLOADS', default=False):
+        shutil.rmtree(YOUTUBE_DOWNLOADS_DIR, ignore_errors=True)
     os.makedirs(YOUTUBE_DOWNLOADS_DIR, exist_ok=True)
     os.makedirs(PLAYER_CACHE_DIR, exist_ok=True)
 
     # Clear debug.log on startup
     open('./debug.log', 'w').close()
 
-    bot.run(log_level=logging.getLevelName(env.str('LOG_LEVEL', default='ERROR')))
+    scheduler_logger = logging.getLogger('apscheduler.executors')
+    scheduler_logger.setLevel(logging.ERROR)
+
+    with ThreadPoolExecutor() as executor:
+        bot.run(log_level=logging.getLevelName(env.str('LOG_LEVEL', default='ERROR')), executor=executor)
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     main()
