@@ -1,6 +1,6 @@
-__title__ = "NeonBot"
-__author__ = "neonspectrum"
-__version__ = "2.3.1"
+__title__ = 'NeonBot'
+__author__ = 'neonspectrum'
+__version__ = '2.3.1'
 
 from neonbot.bot import NeonBot
 
