@@ -3,14 +3,14 @@ from typing import TYPE_CHECKING
 import discord
 from i18n import t
 
-from .embed import Embed
-from .view import View, Button
-from .. import bot
-from ..enums import Repeat, PlayerState
-from ..utils import log
+from neonbot import bot
+from neonbot.classes.embed import Embed
+from neonbot.classes.view import Button, View
+from neonbot.enums import PlayerState, Repeat
+from neonbot.utils import log
 
 if TYPE_CHECKING:
-    from .player import Player
+    from neonbot.classes.player import Player
 
 
 class PlayerControls:
@@ -27,12 +27,12 @@ class PlayerControls:
         else:
             views[0].style = discord.ButtonStyle.secondary
 
-        views[1].disabled = not (0 <= self.player.current_track -1 < len(self.player.track_list))
+        views[1].disabled = not (0 <= self.player.current_track - 1 < len(self.player.track_list))
 
         if self.player.connection and self.player.connection.is_playing():
-            views[2].emoji = "⏸️"
+            views[2].emoji = '⏸️'
         else:
-            views[2].emoji = "▶️"
+            views[2].emoji = '▶️'
 
         views[3].disabled = self.next_disabled = (
             self.player.repeat == Repeat.OFF
@@ -42,13 +42,13 @@ class PlayerControls:
         )
 
         if self.player.repeat == Repeat.OFF:
-            views[4].emoji = "🔁"
+            views[4].emoji = '🔁'
             views[4].style = discord.ButtonStyle.secondary
         elif self.player.repeat == Repeat.SINGLE:
-            views[4].emoji = "🔂"
+            views[4].emoji = '🔂'
             views[4].style = discord.ButtonStyle.primary
         elif self.player.repeat == Repeat.ALL:
-            views[4].emoji = "🔁"
+            views[4].emoji = '🔁'
             views[4].style = discord.ButtonStyle.primary
 
         if self.player.autoplay:
@@ -72,16 +72,18 @@ class PlayerControls:
                 await bot.send_response(interaction, embed=Embed(t('music.cannot_interact')), ephemeral=True)
                 return
 
-        if button.emoji.name == "▶️":  # play
+        if button.emoji.name == '▶️':  # play
             if self.player.connection.is_paused():
                 await self.player.resume(requester=interaction.user)
             else:
                 self.player.current_track = 0
                 await self.player.play()
-        elif button.emoji.name == "⏸️":  # pause
+        elif button.emoji.name == '⏸️':  # pause
             await self.player.pause(requester=interaction.user)
-        elif button.emoji.name == "⏮️":  # prev
-            bot.loop.create_task(send_message(t('music.player_controls_pressed', action='back', user=interaction.user.mention)))
+        elif button.emoji.name == '⏮️':  # prev
+            bot.loop.create_task(
+                send_message(t('music.player_controls_pressed', action='back', user=interaction.user.mention))
+            )
 
             self.player.jump_to_track = self.player.current_track - 1
             self.player.state = PlayerState.JUMPED
@@ -90,8 +92,10 @@ class PlayerControls:
                 self.player.next()
             else:
                 bot.loop.create_task(self.player.after())
-        elif button.emoji.name == "⏭️":  # next
-            bot.loop.create_task(send_message(t('music.player_controls_pressed', action='next', user=interaction.user.mention)))
+        elif button.emoji.name == '⏭️':  # next
+            bot.loop.create_task(
+                send_message(t('music.player_controls_pressed', action='next', user=interaction.user.mention))
+            )
 
             if self.player.connection.is_playing():
                 if self.player.current_track != len(self.player.track_list) - 1:
@@ -101,32 +105,32 @@ class PlayerControls:
                 self.player.next()
             else:
                 bot.loop.create_task(self.player.after())
-        elif button.emoji.name in ("🔁", "🔂"):  # repeat
+        elif button.emoji.name in ('🔁', '🔂'):  # repeat
             modes = [Repeat.OFF, Repeat.SINGLE, Repeat.ALL]
             index = (modes.index(Repeat(self.player.repeat)) + 1) % 3
             await self.player.set_repeat(modes[index], requester=interaction.user)
-        elif button.emoji.name == "🔀":  # shuffle
+        elif button.emoji.name == '🔀':  # shuffle
             await self.player.set_shuffle(requester=interaction.user)
-        elif button.emoji.name == "♾️":  # autoplay
+        elif button.emoji.name == '♾️':  # autoplay
             await self.player.set_autoplay(requester=interaction.user)
-        elif button.emoji.name == "⏹️": # stop
+        elif button.emoji.name == '⏹️':  # stop
             await self.player.stop()
             await send_message(t('music.player_controls_pressed', action='stop', user=interaction.user.mention))
-        elif button.emoji.name == "⏏️": # reset
+        elif button.emoji.name == '⏏️':  # reset
             await self.player.reset()
             self.player.remove_instance()
             await send_message(t('music.player_controls_pressed', action='reset', user=interaction.user.mention))
 
     def initialize(self) -> None:
         buttons = [
-            Button(emoji="🔀"),
-            Button(emoji="⏮️", style=discord.ButtonStyle.primary),
-            Button(emoji="⏸️", style=discord.ButtonStyle.primary),
-            Button(emoji="⏭️", style=discord.ButtonStyle.primary),
-            Button(emoji="🔁"),
-            Button(emoji="♾️", label="Autoplay"),
-            Button(emoji="⏹️", label="Stop"),
-            Button(emoji="⏏️", label="Reset"),
+            Button(emoji='🔀'),
+            Button(emoji='⏮️', style=discord.ButtonStyle.primary),
+            Button(emoji='⏸️', style=discord.ButtonStyle.primary),
+            Button(emoji='⏭️', style=discord.ButtonStyle.primary),
+            Button(emoji='🔁'),
+            Button(emoji='♾️', label='Autoplay'),
+            Button(emoji='⏹️', label='Stop'),
+            Button(emoji='⏏️', label='Reset'),
         ]
         self.update_buttons(buttons)
 
