@@ -1,12 +1,12 @@
 import os
 import random
 from datetime import datetime
-from io import BytesIO
 from time import time
-from typing import cast
+from typing import Union, cast
 
 import discord
 import psutil
+from dateutil.parser import parse
 from discord import app_commands
 from discord.ext import commands
 from discord.utils import format_dt
@@ -17,7 +17,7 @@ from neonbot import __author__, __title__, __version__, bot
 from neonbot.classes.embed import Embed
 from neonbot.classes.gemini import GeminiChat
 from neonbot.utils.constants import ICONS
-from neonbot.utils.functions import format_seconds, md_to_text
+from neonbot.utils.functions import format_seconds, generate_profile_member_embed, generate_profile_user_embed
 
 
 class Utility(commands.Cog):
@@ -29,7 +29,8 @@ class Utility(commands.Cog):
         """Picks a text in the given list."""
 
         await cast(discord.InteractionResponse, interaction.response).send_message(
-            embed=Embed(random.choice(word_list.split(',')).strip()))
+            embed=Embed(random.choice(word_list.split(',')).strip())
+        )
 
     @app_commands.command(name='stats')
     @app_commands.allowed_installs(guilds=True, users=True)
@@ -40,27 +41,25 @@ class Utility(commands.Cog):
         process = psutil.Process(os.getpid())
 
         embed = Embed()
-        embed.set_author(f"{__title__} v{__version__}", icon_url=bot.user.display_avatar)
-        embed.add_field("Username", bot.user.name)
-        embed.add_field("Created On", f"{bot.user.created_at:%Y-%m-%d %I:%M:%S %p}")
-        embed.add_field("Created By", __author__)
-        embed.add_field("Guilds", len(bot.guilds))
-        embed.add_field("Channels", sum(1 for _ in bot.get_all_channels()))
-        embed.add_field("Users", len(bot.users))
+        embed.set_author(f'{__title__} v{__version__}', icon_url=bot.user.display_avatar)
+        embed.add_field('Username', bot.user.name)
+        embed.add_field('Created On', f'{bot.user.created_at:%Y-%m-%d %I:%M:%S %p}')
+        embed.add_field('Created By', __author__)
+        embed.add_field('Guilds', len(bot.guilds))
+        embed.add_field('Channels', sum(1 for _ in bot.get_all_channels()))
+        embed.add_field('Users', len(bot.users))
         embed.add_field(
-            "Ram Usage",
-            f"Approximately {(process.memory_info().rss / 1024000):.2f} MB",
+            'Ram Usage',
+            f'Approximately {(process.memory_info().rss / 1024000):.2f} MB',
             inline=True,
         )
+        embed.add_field('Uptime', format_seconds(time() - process.create_time()).split('.')[0])
         embed.add_field(
-            "Uptime", format_seconds(time() - process.create_time()).split(".")[0]
-        )
-        embed.add_field(
-            "Packages",
+            'Packages',
             f"""
             discord `{discord.__version__}`
             youtube-dl `{ytdl_version.__version__}`
-            """
+            """,
         )
 
         await cast(discord.InteractionResponse, interaction.response).send_message(embed=embed)
@@ -72,46 +71,41 @@ class Utility(commands.Cog):
 
         def generate_embed():
             embed = Embed()
-            embed.set_author(name="✉ SMS")
-            embed.set_footer(
-                text="Powered by Semaphore",
-                icon_url=ICONS['semaphone']
-            )
-            embed.add_field("To:", number, inline=True)
-            embed.add_field("Body:", message, inline=True)
+            embed.set_author(name='✉ SMS')
+            embed.set_footer(text='Powered by Semaphore', icon_url=ICONS['semaphone'])
+            embed.add_field('To:', number, inline=True)
+            embed.add_field('Body:', message, inline=True)
 
             return embed
 
         await cast(discord.InteractionResponse, interaction.response).send_message(
-            embed=generate_embed().add_field("Status:", "Sending...", inline=False))
+            embed=generate_embed().add_field('Status:', 'Sending...', inline=False)
+        )
 
-        api_key = env.str("SEMAPHONE_API_KEY")
-        sender_name = env.str("SEMAPHONE_SENDER_NAME")
+        api_key = env.str('SEMAPHONE_API_KEY')
+        sender_name = env.str('SEMAPHONE_SENDER_NAME')
 
-        body = f"{message}\n\nSent by {interaction.user}"
+        body = f'{message}\n\nSent by {interaction.user}'
 
         response = await bot.session.post(
-            f"https://api.semaphore.co/api/v4/messages",
-            data={
-                "apikey": api_key,
-                "sendername": sender_name,
-                "message": body,
-                "number": number
-            }
+            'https://api.semaphore.co/api/v4/messages',
+            data={'apikey': api_key, 'sendername': sender_name, 'message': body, 'number': number},
         )
 
         if response.status >= 400:
             data = await response.json()
 
             await interaction.edit_original_response(
-                embed=generate_embed().add_field("Status:", "Sending failed.", inline=False)
-                .add_field("Reason:", data['status'], inline=False)
-                .add_field("Date sent:", format_dt(datetime.now()), inline=False)
+                embed=generate_embed()
+                .add_field('Status:', 'Sending failed.', inline=False)
+                .add_field('Reason:', data['status'], inline=False)
+                .add_field('Date sent:', format_dt(datetime.now()), inline=False)
             )
         else:
             await interaction.edit_original_response(
-                embed=generate_embed().add_field("Status:", "Sent", inline=False)
-                .add_field("Date sent:", format_dt(datetime.now()), inline=False)
+                embed=generate_embed()
+                .add_field('Status:', 'Sent', inline=False)
+                .add_field('Date sent:', format_dt(datetime.now()), inline=False)
             )
 
     @app_commands.command(name='ask')
@@ -128,6 +122,63 @@ class Utility(commands.Cog):
         await gemini_chat.generate_content()
 
         await interaction.followup.send(gemini_chat.get_response())
+
+    @app_commands.command(name='profile')
+    @app_commands.allowed_installs(guilds=True, users=True)
+    @app_commands.allowed_contexts(guilds=True, dms=True, private_channels=True)
+    async def profile(
+        self, interaction: discord.Interaction, user: Union[discord.User, discord.Member], ephemeral: bool = False
+    ) -> None:
+        if isinstance(user, discord.Member):
+            embed = await generate_profile_member_embed(interaction, user)
+        else:
+            embed = await generate_profile_user_embed(interaction, user)
+        await cast(discord.InteractionResponse, interaction.response).send_message(embed=embed, ephemeral=ephemeral)
+
+    @app_commands.command(name='avatar')
+    @app_commands.allowed_installs(guilds=True, users=True)
+    @app_commands.allowed_contexts(guilds=True, dms=True, private_channels=True)
+    async def avatar(
+        self, interaction: discord.Interaction, user: Union[discord.User, discord.Member], ephemeral: bool = False
+    ) -> None:
+        if not user.display_avatar:
+            await cast(discord.InteractionResponse, interaction.response).send_message(
+                embed=Embed('Avatar not found.'), ephemeral=True
+            )
+            return
+
+        await cast(discord.InteractionResponse, interaction.response).send_message(
+            user.display_avatar.url, ephemeral=ephemeral
+        )
+
+    @app_commands.command(name='banner')
+    @app_commands.allowed_installs(guilds=True, users=True)
+    @app_commands.allowed_contexts(guilds=True, dms=True, private_channels=True)
+    async def banner(
+        self, interaction: discord.Interaction, user: Union[discord.User, discord.Member], ephemeral: bool = False
+    ) -> None:
+        user = await bot.fetch_user(user.id)
+
+        if not user.banner:
+            await cast(discord.InteractionResponse, interaction.response).send_message(
+                embed=Embed('Banner not found.'), ephemeral=True
+            )
+            return
+
+        await cast(discord.InteractionResponse, interaction.response).send_message(user.banner.url, ephemeral=ephemeral)
+
+    @app_commands.command(name='timestamp')
+    @app_commands.allowed_installs(guilds=True, users=True)
+    @app_commands.allowed_contexts(guilds=True, dms=True, private_channels=True)
+    async def timestamp(
+        self, interaction: discord.Interaction, date_string: str
+    ) -> None:
+        try:
+            timestamp = int(parse(date_string).timestamp())
+            await cast(discord.InteractionResponse, interaction.response).send_message(f'<t:{timestamp}>')
+        except ValueError:
+            await cast(discord.InteractionResponse, interaction.response).send_message('Format not supported.',
+                                                                                   ephemeral=True)
 
 
 # noinspection PyShadowingNames
