@@ -1,2 +1,2 @@
-from .player_state import PlayerState
-from .repeat import Repeat
+from neonbot.enums.player_state import PlayerState
+from neonbot.enums.repeat import Repeat
