@@ -1,9 +1,7 @@
-from typing import Optional
-
 from pydantic import BaseModel
 
 
-class Music(BaseModel):
+class MusicModel(BaseModel):
     volume: int
     repeat: int
     shuffle: bool
