@@ -28,4 +28,5 @@ class WishlistModal(discord.ui.Modal, title='Set your wishlist'):
             await self.parent.edit_original_response(embed=Embed(f'Your wishlist: ```{wishlist}```'))
         else:
             await cast(discord.InteractionResponse, interaction.response).send_message(
-                embed=Embed('Your wishlist has been updated.'), ephemeral=True)
+                embed=Embed('Your wishlist has been updated.'), ephemeral=True
+            )
